# backend/core/api-specs.yaml
openapi: 3.0.3
info:
  title: Tattara DSN AI Data Collection App API
  description: |
    Backend API for the Tattara DSN AI Data Collection App.
    Features:
      - JWT Authentication with RBAC (Roles & Permissions)
      - Program & Workflow management
      - AI-assisted data collection (Voice, OCR) with user confirmation step
      - Dual integration: DHIS2 and Generic Database
      - Offline sync & comprehensive audit logging
  version: 2.0.0
servers:
  - url: https://api.tattara.app/v1
...
security:
  - BearerAuth: []
paths:
...
  /workflows:
    get:
      tags: [Workflow Management]
      summary: List all workflows
      parameters:
        - $ref: '#/components/parameters/ProgramIdQueryParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowListResponse'
    post:
      tags: [Workflow Management]
      summary: Create a new workflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkflowRequest'
      responses:
        '201':
          description: Workflow created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
  /workflows/{workflowId}/configurations:
    get:
      tags: [Workflow Management]
      summary: Get all external configurations for a workflow
      parameters:
        - $ref: '#/components/parameters/WorkflowIdPathParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowConfigurationListResponse'
    post:
      tags: [Workflow Management]
      summary: Add a target configuration to a workflow
      parameters:
        - $ref: '#/components/parameters/WorkflowIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkflowConfigRequest'
      responses:
        '201':
          $ref: '#/components/responses/Success'
  /workflows/{workflowId}/fields:
    get:
      tags: [Workflow Management]
      summary: Get all fields for a workflow
      parameters:
        - $ref: '#/components/parameters/WorkflowIdPathParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowFieldListResponse'
    post:
      tags: [Workflow Management]
      summary: Add a field to a workflow
      parameters:
        - $ref: '#/components/parameters/WorkflowIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkflowFieldRequest'
      responses:
        '201':
          $ref: '#/components/responses/Success'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    WorkflowIdPathParam:
      name: workflowId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/UUID' }
...
    ProgramIdQueryParam:
      name: programId
      in: query
      schema: { $ref: '#/components/schemas/UUID' }
...
  responses:
    Success:
      description: Operation successful
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessResponse'
  schemas:
    UUID:
      type: string
      format: uuid
      example: '123e4567-e89b-12d3-a456-426614174000'

    # AUTH
...
    Workflow:
      type: object
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        name: { type: string }
        description: { type: string }
        status: { type: string, enum: [active, inactive, archived] }
        programId: { $ref: '#/components/schemas/UUID' }
        fields:
          { type: array, items: { $ref: '#/components/schemas/WorkflowField' } }
        configurations:
          {
            type: array,
            items: { $ref: '#/components/schemas/WorkflowConfiguration' },
          }
    CreateWorkflowRequest:
      type: object
      required: [name, programId]
      properties:
        name: { type: string }
        description: { type: string }
        programId: { $ref: '#/components/schemas/UUID' }
    WorkflowListResponse:
      type: object
      properties:
        workflows:
          { type: array, items: { $ref: '#/components/schemas/Workflow' } }
        pagination: { $ref: '#/components/schemas/PaginationInfo' }
    WorkflowField:
      type: object
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        fieldId: { type: string }
        label: { type: string }
        type:
          { type: string, enum: [text, number, date, select, boolean, etc.] }
        isRequired: { type: boolean }
        validationRules: { type: object }
    CreateWorkflowFieldRequest:
      type: object
      required: [fieldId, label, type]
      properties:
        fieldId: { type: string }
        label: { type: string }
        type: { type: string }
        isRequired: { type: boolean, default: false }
        validationRules: { type: object }
    WorkflowFieldListResponse:
      type: object
      properties:
        fields:
          { type: array, items: { $ref: '#/components/schemas/WorkflowField' } }
    WorkflowConfiguration:
      type: object
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        type: { type: string, enum: [dhis2, postgresql, mysql, mssql, api] }
        isActive: { type: boolean }
        configuration: { type: object } # Discriminated union based on 'type'
    CreateWorkflowConfigRequest:
      type: object
      required: [type, configuration]
      properties:
        type: { type: string }
        configuration: { type: object }
    WorkflowConfigurationListResponse:
      type: object
      properties:
        configurations:
          {
            type: array,
            items: { $ref: '#/components/schemas/WorkflowConfiguration' },
          }
...
    SuccessResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string }
    PaginationInfo:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }